promote experimental dashboard to default and remove experimental mode
- route `/pages/home` directly to `HomeComponent` and rename former experimental files/symbols
- remove dashboard toggle from Edit page and delete obsolete toggle service/shell/legacy home paths
- remove `experimental` history gate; apply `history_span` directly (with existing limits)
- update Home docs/comments to default-dashboard terminology
