+++ main/http_server/axe-os/src/app/pages/edit/edit.component.html
@@ -366,9 +366,6 @@
             <nb-card-header>{{ 'MISC.TITLE' | translate }}</nb-card-header>
             <nb-card-body>
                 <div class="form-row">
-                  <nb-checkbox formControlName="experimentalDashboardEnabled">Use EXPERIMENTAL Dashboard</nb-checkbox>
-                </div>
-                <div class="form-row">
                     <nb-checkbox formControlName="flipscreen">{{ 'MISC.FLIP_SCREEN' | translate }}</nb-checkbox>
                 </div>
                 <div class="form-row">
